# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RoutineOps Server API
  version: 2.4.6
  description: |
    HTTP API сервера RoutineOps — точка интеграции для автоматизации и сторонних
    интерфейсов. Веб-интерфейс из поставки использует этот же API, отдельного
    приватного канала у него нет.

    ## Аутентификация

    Два равноправных способа, оба проверяются одним middleware:

    * **Cookie `token`** — httpOnly-кука, выставляется `POST /api/v1/auth/login`.
      Так работает веб-интерфейс.
    * **`Authorization: Bearer <token>`** — сервисный API-токен для автоматизации.
      Выпускается через `POST /api/v1/api-tokens`, показывается ОДИН раз в ответе.

    ## Роли

    * `viewer` — чтение.
    * `it_admin` — чтение и изменение.

    ## Операции «только человеком»

    Часть операций помечена `x-requires-human: true`. Они доступны только под
    учётной записью человека и отвечают `403` на сервисный токен, даже если у
    токена роль `it_admin`. Правило распространяется на всё, что выпускает или
    повышает права: приглашение пользователя, выпуск API-токена и токена
    энроллмента, одобрение устройства и запроса локальных админ-прав, вывод
    устройства из эксплуатации.

    Смысл в том, что иначе модель отзыва была бы фикцией: утёкший токен выписал
    бы себе новый, который пережил бы удаление исходного.

    ## Ошибки

    Тело ошибки — `text/plain` с короткой строкой, не JSON. Значим код ответа.

  license:
    name: Apache-2.0
...
servers:
  - url: https://mdm.example.com
...
security:
  - cookieAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/oidc/providers:
    x-enterprise: true
    get:
      tags: [Аутентификация]
      summary: Провайдеры SSO по OIDC
      description: |
        Enterprise. Роль `it_admin`, только человеком: кто заводит IdP, тот заводит
        вход, а это выпуск прав.

        `client_secret` наружу не отдаётся — только признак `has_secret`.
      x-requires-human: true
      responses:
        '200':
          description: Список провайдеров тенанта
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/OIDCProvider' }
        '403': { $ref: '#/components/responses/ForbiddenHuman' }
        '501': { description: Сборка без поддержки SSO (open-core) }
    post:
      tags: [Аутентификация]
      summary: Добавить провайдер OIDC
      description: |
        `issuer_url` обязан быть `https` (http допустим только на loopback): по нему
        едут discovery и JWKS, то есть ключи проверки подписи `id_token`.
      x-requires-human: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OIDCProviderInput' }
      responses:
        '201':
          description: Создан
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OIDCProvider' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/ForbiddenHuman' }
        '501': { description: Сборка без поддержки SSO (open-core) }
  /api/v1/oidc/providers/{id}:
    x-enterprise: true
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    put:
      tags: [Аутентификация]
      summary: Изменить провайдер OIDC
      description: Пустой `client_secret` означает «не менять».
      x-requires-human: true
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OIDCProviderInput' }
      responses:
        '204': { description: Обновлён }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/ForbiddenHuman' }
        '404': { $ref: '#/components/responses/NotFound' }
        '501': { description: Сборка без поддержки SSO (open-core) }
    delete:
      tags: [Аутентификация]
      summary: Удалить провайдер OIDC
      x-requires-human: true
      responses:
        '204': { description: Удалён }
        '403': { $ref: '#/components/responses/ForbiddenHuman' }
        '404': { $ref: '#/components/responses/NotFound' }
        '501': { description: Сборка без поддержки SSO (open-core) }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: httpOnly-кука, выставляется входом. Так работает веб-интерфейс.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Сервисный API-токен для автоматизации.
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content: { text/plain: { schema: { type: string } } }
...
    ForbiddenHuman:
      description: |
        Недостаточно прав. Операция доступна только человеку: сервисный токен
        получает этот ответ даже с ролью `it_admin`.
      content: { text/plain: { schema: { type: string } } }
    NotFound:
      description: Не найдено
      content: { text/plain: { schema: { type: string } } }
...
  schemas:
...
    OIDCProvider:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        client_id: { type: string }
        issuer_url: { type: string }
        redirect_uri: { type: string }
        enabled: { type: boolean }
        has_secret: { type: boolean, description: client_secret задан; сам секрет не отдаётся }
    OIDCProviderInput:
      type: object
      required: [name, client_id, issuer_url, redirect_uri]
      properties:
        name: { type: string }
        client_id: { type: string }
        client_secret: { type: string, description: При обновлении пусто = не менять }
        issuer_url: { type: string, description: Обязательно https (http только на loopback) }
        redirect_uri: { type: string }
        enabled: { type: boolean }
...